Cleaning/Formatting

+++ std/src/Rmultiply.cpp
@@ -101,10 +101,10 @@ ResultSegment* Rmultiply::renderSegment(ResultSegmentSettings* resSettings, Rend
 
 			result = new Dbimg(*fmt);
 
-			const int height = result->getHeight();
-			const int width = result->getWidth();
-			const int channels = 4;
-			const long dataSize = height*width*channels;
+			const uint32_t height = result->getHeight();
+			const uint32_t width = result->getWidth();
+			const uint32_t channels = 4;
+			const size_t dataSize = height*width*channels;
 
 
 			uint8_t* srcA = a.getResult()->getImageData();
